requirements.txt
# Pneumonia Detection - Streamlit Cloud Requirements

# Core ML libraries (CPU version for Streamlit Cloud)
tensorflow-cpu==2.13.0
numpy==1.24.3
pillow==10.0.0
opencv-python-headless==4.8.0.76

# Web Framework
streamlit==1.28.0
streamlit-folium==0.15.0

# Data processing
pandas==2.0.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Maps and Location
folium==0.15.0

# API and Web
requests==2.31.0

# Utilities
python-dotenv==1.0.0